refactor: 페이지 내에 스타일 외부 파일로 이동
