+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import styled from 'styled-components';
-import AuthPage from './page/AuthPage';
-import BookmarkPage from './page/BookmarkPage';
-import DetailPage from './page/DetailPage';
-import FindPage from './page/FindPage';
-import MainPage from './page/MainPage';
+import AuthPage from './page/AuthPage/AuthPage';
+import BookmarkPage from './page/BookmarkPage/BookmarkPage';
+import DetailPage from './page/DetailPage/DetailPage';
+import FindPage from './page/FindPage/FindPage';
+import MainPage from './page/MainPage/MainPage';
 
 function App() {
 	return (
